server/rest_api_in_django/requirements.txt
django
djangorestframework
pillow
# opencv-python (optional)